+++ lib/core/src/data.rs
@@ -1,4 +1,5 @@
 pub mod container;
+pub mod dirty;
 pub mod function;
 pub mod opt_vec;
 pub mod seq;
