+++ lib/core/src/data/seq.rs
