+++ lib/core/src/display/symbol.rs
@@ -7,8 +7,8 @@ use crate::backend::webgl;
 use crate::backend::webgl::Context;
 use crate::closure;
 use crate::data::function::callback::*;
-use crate::dirty;
-use crate::dirty::traits::*;
+use crate::data::dirty;
+use crate::data::dirty::traits::*;
 use crate::display::symbol::geometry::primitive::mesh;
 use crate::promote;
 use crate::promote_all;
