+++ lib/core/src/display/symbol/geometry/primitive/mesh/buffer.rs
@@ -6,8 +6,8 @@ use crate::prelude::*;
 use crate::backend::webgl::Context;
 use crate::closure;
 use crate::data::function::callback::*;
-use crate::dirty;
-use crate::dirty::traits::*;
+use crate::data::dirty;
+use crate::data::dirty::traits::*;
 use crate::system::web::fmt;
 use crate::system::web::group;
 use crate::system::web::Logger;
