+++ lib/core/src/display/world.rs
@@ -10,8 +10,8 @@ use crate::closure;
 use crate::control::callback::CallbackHandle;
 use crate::control::event_loop::EventLoop;
 use crate::data::opt_vec::OptVec;
-use crate::dirty;
-use crate::dirty::traits::*;
+use crate::data::dirty;
+use crate::data::dirty::traits::*;
 use crate::promote_all;
 use crate::promote_workspace_types;
 use crate::promote;
