+++ lib/core/src/display/world/scene/mesh_registry.rs
@@ -4,8 +4,8 @@ use crate::backend::webgl::Context;
 use crate::closure;
 use crate::data::function::callback::*;
 use crate::data::opt_vec::OptVec;
-use crate::dirty;
-use crate::dirty::traits::*;
+use crate::data::dirty;
+use crate::data::dirty::traits::*;
 use crate::display::symbol;
 use crate::system::web::group;
 use crate::system::web::Logger;
