+++ lib/core/src/display/world/workspace.rs
@@ -5,8 +5,8 @@ pub use crate::display::world::scene::mesh_registry::ObjectId;
 use crate::backend::webgl;
 use crate::closure;
 use crate::data::function::callback::*;
-use crate::dirty;
-use crate::dirty::traits::*;
+use crate::data::dirty;
+use crate::data::dirty::traits::*;
 use crate::display::world::scene::mesh_registry;
 use crate::display::world::scene::Scene;
 use crate::promote_all;
